+++ paddle/fluid/operators/fused/attn_bias_add.cu.h
@@ -191,10 +191,9 @@ void SetConfigForColumnReduce(const int max_threads, const int reduce_num,
 
   int num_block = (max_threads / left_num);
   if (num_block > 1 && reduce_num >= REDUCE_SPLIT_BOUNDARY) {
-    *blocking_size =
-        pten::kernels::details::GetLastPow2(reduce_num / num_block);
+    *blocking_size = pten::funcs::details::GetLastPow2(reduce_num / num_block);
     if (*blocking_size <= 1) {
-      *blocking_size = pten::kernels::details::GetLastPow2(sqrt(reduce_num));
+      *blocking_size = pten::funcs::details::GetLastPow2(sqrt(reduce_num));
     } else if (*blocking_size * 2 < reduce_num) {
       *blocking_size *= 2;
     }
