+++ paddle/fluid/operators/reduce_ops/reduce_op.cu.h
@@ -23,7 +23,7 @@
 #include "paddle/fluid/framework/tensor.h"
 
 #include "paddle/pten/core/dense_tensor.h"
-#include "paddle/pten/kernels/gpu/reduce.h"
+#include "paddle/pten/kernels/funcs/reduce_function.h"
 
 namespace paddle {
 namespace operators {
@@ -37,7 +37,7 @@ void TensorReduceImpl(const platform::CUDADeviceContext& dev_ctx,
                       gpuStream_t stream) {
   y->mutable_data<Ty>(x.place());
 
-  pten::kernels::TensorReduceImpl<Tx, Ty, ReduceOp, TransformOp>(
+  pten::funcs::TensorReduceImpl<Tx, Ty, ReduceOp, TransformOp>(
       static_cast<const pten::GPUContext&>(dev_ctx), x, y, transform,
       origin_reduce_dims, stream);
 }
